+++ src/transpile.rs
@@ -560,7 +560,7 @@ impl FunctionCallBuilder {
         let func_name = quote_ident(&self.function.name);
 
         let query = match &self.return_type_builder {
-            FuncCallReturnTypeBuilder::Scalar => {
+            FuncCallReturnTypeBuilder::Scalar | FuncCallReturnTypeBuilder::List => {
                 let type_adjustment_clause = apply_suffix_casts(self.function.type_oid);
                 format!("select to_jsonb({func_schema}.{func_name}{args_clause}{type_adjustment_clause}) {block_name};")
             }
@@ -1353,6 +1353,9 @@ fn apply_suffix_casts(type_oid: u32) -> String {
         20 => "::text",           // bigints as text
         114 | 3802 => "#>> '{}'", // json/b as stringified
         1700 => "::text",         // numeric as text
+        1016 => "::text[]",       // bigint arrays as array of text
+        199 | 3807 => "::text[]", // json/b array as array of text
+        1231 => "::text[]",       // numeric array as array of text
         _ => "",
     }
     .to_string()