+++ test/sql/function_calls_unsupported.sql
@@ -107,26 +107,6 @@ begin;
         }
     $$));
 
-    create function func_accepting_array(a int[])
-        returns int language sql immutable
-    as $$ select 0; $$;
-
-    select jsonb_pretty(graphql.resolve($$
-        query {
-            funcAcceptingArray(a: [1, 2, 3])
-        }
-    $$));
-
-    create function func_returning_array()
-        returns int[] language sql immutable
-    as $$ select array[1, 2, 3]; $$;
-
-    select jsonb_pretty(graphql.resolve($$
-        query {
-            funcReturningArray
-        }
-    $$));
-
     -- function returning type not on search path
     create schema dev;
     create table dev.book(
@@ -162,6 +142,31 @@ begin;
     select jsonb_pretty(graphql.resolve($$
         query {
             badInputArg
+        }
+    $$));
+
+
+    -- function returning enum
+    create type "Algorithm" as enum ('aead-ietf');
+    comment on type "Algorithm" is '@graphql({"mappings": {"aead-ietf": "AEAD_IETF"}})';
+
+    create function return_algorithm()
+        returns "Algorithm" language sql volatile
+    as $$ select 'aead-ietf'::"Algorithm"; $$;
+
+    select jsonb_pretty(graphql.resolve($$
+        mutation {
+            returnAlgorithm
+        }
+    $$));
+
+    create function accept_algorithm(e "Algorithm")
+        returns int language sql stable
+    as $$ select 0; $$;
+
+    select jsonb_pretty(graphql.resolve($$
+        query {
+            acceptAlgorithm(e: "AEAD_IETF")
         }
     $$));